refactor(materializer): extract operation-patch conversation strategies (file #3)
Lift materializeOperationPatchedConversation + materializeBlockingInteractionActivityChange
into operation-patch-conversations.ts; dispatcher imports them. Net 74/74, check 0.
